# [common] Speed up ConcurrentHashMap#computeIfAbsent