--- a/fluss-rpc/src/main/java/com/alibaba/fluss/rpc/metrics/ConnectionMetricGroup.java
+++ b/fluss-rpc/src/main/java/com/alibaba/fluss/rpc/metrics/ConnectionMetricGroup.java
@@ -25,13 +25,13 @@ import com.alibaba.fluss.metrics.groups.AbstractMetricGroup;
 import com.alibaba.fluss.metrics.groups.MetricGroup;
 import com.alibaba.fluss.metrics.registry.MetricRegistry;
 import com.alibaba.fluss.rpc.protocol.ApiKeys;
+import com.alibaba.fluss.utils.concurrent.ConcurrentHashMapUtils;
 
 import javax.annotation.Nullable;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static com.alibaba.fluss.metrics.utils.MetricGroupUtils.makeScope;
@@ -44,7 +44,8 @@ public class ConnectionMetricGroup extends AbstractMetricGroup {
     private final String serverId;
 
     /** Metrics for different request/response metrics with specify {@link ApiKeys}. */
-    private final Map<String, Metrics> metricsByRequestName = new ConcurrentHashMap<>();
+    private final Map<String, Metrics> metricsByRequestName =
+            ConcurrentHashMapUtils.newConcurrentHashMap();
 
     public ConnectionMetricGroup(
             MetricRegistry registry, String serverId, ClientMetricGroup parent) {