--- a/fluss-server/src/main/java/com/alibaba/fluss/server/metrics/group/TabletServerMetricGroup.java
+++ b/fluss-server/src/main/java/com/alibaba/fluss/server/metrics/group/TabletServerMetricGroup.java
@@ -24,9 +24,9 @@ import com.alibaba.fluss.metrics.MetricNames;
 import com.alibaba.fluss.metrics.ThreadSafeSimpleCounter;
 import com.alibaba.fluss.metrics.groups.AbstractMetricGroup;
 import com.alibaba.fluss.metrics.registry.MetricRegistry;
+import com.alibaba.fluss.utils.concurrent.ConcurrentHashMapUtils;
 
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 /** The metric group for tablet server. */
 public class TabletServerMetricGroup extends AbstractMetricGroup {
@@ -34,7 +34,7 @@ public class TabletServerMetricGroup extends AbstractMetricGroup {
     private static final String NAME = "tabletserver";
 
     private final Map<PhysicalTablePath, PhysicalTableMetricGroup> metricGroupByPhysicalTable =
-            new ConcurrentHashMap<>();
+            ConcurrentHashMapUtils.newConcurrentHashMap();
 
     protected final String clusterId;
     protected final String hostname;