--- a/fluss-server/src/main/java/com/alibaba/fluss/server/replica/AdjustIsrManager.java
+++ b/fluss-server/src/main/java/com/alibaba/fluss/server/replica/AdjustIsrManager.java
@@ -26,6 +26,7 @@ import com.alibaba.fluss.rpc.protocol.Errors;
 import com.alibaba.fluss.server.entity.AdjustIsrResultForBucket;
 import com.alibaba.fluss.server.utils.RpcMessageUtils;
 import com.alibaba.fluss.server.zk.data.LeaderAndIsr;
+import com.alibaba.fluss.utils.concurrent.ConcurrentHashMapUtils;
 import com.alibaba.fluss.utils.concurrent.Scheduler;
 
 import org.slf4j.Logger;
@@ -36,7 +37,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /* This file is based on source code of Apache Kafka Project (https://kafka.apache.org/), licensed by the Apache
@@ -59,7 +59,8 @@ public class AdjustIsrManager {
     private final int serverId;
 
     /** Used to allow only one pending adjust Isr request per bucket (visible for testing). */
-    protected final Map<TableBucket, AdjustIsrItem> unsentAdjustIsrMap = new ConcurrentHashMap<>();
+    protected final Map<TableBucket, AdjustIsrItem> unsentAdjustIsrMap =
+            ConcurrentHashMapUtils.newConcurrentHashMap();
 
     /** Used to allow only one in-flight request at a time. */
     private final AtomicBoolean inflightRequest = new AtomicBoolean(false);