--- a/fluss-server/src/main/java/com/alibaba/fluss/server/replica/delay/DelayedOperationManager.java
+++ b/fluss-server/src/main/java/com/alibaba/fluss/server/replica/delay/DelayedOperationManager.java
@@ -22,6 +22,7 @@ import com.alibaba.fluss.server.utils.timer.DefaultTimer;
 import com.alibaba.fluss.server.utils.timer.Timer;
 import com.alibaba.fluss.server.utils.timer.TimerTask;
 import com.alibaba.fluss.utils.Preconditions;
+import com.alibaba.fluss.utils.concurrent.ConcurrentHashMapUtils;
 import com.alibaba.fluss.utils.concurrent.ShutdownableThread;
 
 import org.slf4j.Logger;
@@ -33,7 +34,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Queue;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.Lock;
@@ -235,7 +235,7 @@ public final class DelayedOperationManager<T extends DelayedOperation> {
         private final Lock watcherLock;
 
         public WatcherList() {
-            this.watchersByKey = new ConcurrentHashMap<>();
+            this.watchersByKey = ConcurrentHashMapUtils.newConcurrentHashMap();
             this.watcherLock = new ReentrantLock();
         }